chore: moved listener code, removed unusued form
