+++ index.html
@@ -17,7 +17,6 @@
 
 		<!-- Scripts -->
 		<script defer type="module" src="/src/main.ts"></script>
-		<script defer type="module" src="/src/listeners.ts"></script>
 	</head>
 	<body class="relative h-screen w-screen">
 		<dialog id="donate-dialog">
