+++ src/components/donate-modal.ts
@@ -21,17 +21,12 @@ export class DonateModal extends LitElement {
 				.isOpen=${this.isOpen}
 				@modal-closed=${() => this.handleModalClosed()}
 			>
-				<form
-					class="relative rounded-xl bg-white p-8"
-					method="dialog"
+				<div
 					slot="content"
-				>
-					<div
-						id="mmd_widget_iframe_div"
-						data-dedication-value="Anonymous Sapling"
-						data-topic="tree_planting"
-					></div>
-				</form>
+					id="mmd_widget_iframe_div"
+					data-dedication-value="Anonymous Sapling"
+					data-topic="tree_planting"
+				></div>
 			</base-modal>
 		`;
 	}
